refactor: extract refining steps to widget form store
