+++ apps/web/src/services/dashboards/widgets/_composables/use-widget/merge-base-widget-state.ts
@@ -1,8 +1,6 @@
 import type { ComputedRef, Ref, UnwrapRef } from 'vue';
 import { computed, reactive } from 'vue';
 
-import { union } from 'lodash';
-
 import type {
     DashboardSettings,
     DashboardVariables,
@@ -10,13 +8,11 @@ import type {
 } from '@/schema/dashboard/_types/dashboard-type';
 import type { InheritOptions, WidgetConfig, WidgetOptions } from '@/schema/dashboard/_types/widget-type';
 
-import type { DashboardScope } from '@/services/dashboards/types/dashboard-view-type';
 import { getWidgetConfig } from '@/services/dashboards/widgets/_helpers/widget-config-helper';
 import { getInitialWidgetInheritOptions } from '@/services/dashboards/widgets/_helpers/widget-inherit-options-helper';
 import {
     getRefinedWidgetOptions,
 } from '@/services/dashboards/widgets/_helpers/widget-options-helper';
-import { getWidgetOptionsSchema } from '@/services/dashboards/widgets/_helpers/widget-options-schema-generator';
 import {
     getInitialSchemaProperties, getRefinedSchemaProperties,
 } from '@/services/dashboards/widgets/_helpers/widget-schema-helper';
@@ -40,11 +36,10 @@ interface MergeBaseWidgetStateOptions {
     dashboardVariables: DashboardVariables|undefined|Ref<DashboardVariables|undefined>; // dashboard variables
     title?: string|Ref<string|undefined>; // widget title from the dashboard widget layout info.
     schemaProperties?: string[]|Ref<string[]|undefined>; // widget schema properties from the dashboard widget layout info.
-    dashboardScope?: DashboardScope|Ref<DashboardScope|undefined>; // it is used for attaching extra options to the widget options, schema properties, and inherit options. default: 'WORKSPACE'
 }
 export function mergeBaseWidgetState(
     {
-        inheritOptions, widgetOptions, widgetName, dashboardSettings, dashboardVariablesSchema, dashboardVariables, title, schemaProperties, dashboardScope,
+        inheritOptions, widgetOptions, widgetName, dashboardSettings, dashboardVariablesSchema, dashboardVariables, title, schemaProperties,
     }: MergeBaseWidgetStateOptions,
 ) {
     const optionState = reactive({
@@ -56,7 +51,6 @@ export function mergeBaseWidgetState(
         dashboardVariables,
         title,
         schemaProperties,
-        dashboardScope,
     });
     const optionsErrorMap = computed(() => getWidgetInheritOptionsErrorMap(
         optionState.schemaProperties ?? [],
@@ -66,10 +60,7 @@ export function mergeBaseWidgetState(
     ));
 
     const state = reactive<MergedBaseWidgetState>({
-        widgetConfig: computed<WidgetConfig>(() => {
-            const config = getWidgetConfig(optionState.widgetName);
-            return getWidgetConfigByDashboardScope(config, optionState.dashboardScope ?? 'WORKSPACE');
-        }),
+        widgetConfig: computed<WidgetConfig>(() => getWidgetConfig(optionState.widgetName)),
         options: computed<WidgetOptions>(() => getRefinedWidgetOptions(
             state.widgetConfig,
             optionState.widgetOptions,
@@ -77,10 +68,7 @@ export function mergeBaseWidgetState(
             optionState.dashboardVariables,
             optionsErrorMap.value,
         )),
-        inheritOptions: computed<InheritOptions>(() => {
-            const _inheritOptions = getInitialWidgetInheritOptions(state.widgetConfig, optionState.inheritOptions, optionState.dashboardVariablesSchema);
-            return getInheritOptionsByDashboardScope(_inheritOptions, optionState.dashboardScope ?? 'WORKSPACE');
-        }),
+        inheritOptions: computed<InheritOptions>(() => getInitialWidgetInheritOptions(state.widgetConfig, optionState.inheritOptions, optionState.dashboardVariablesSchema)),
         settings: computed<DashboardSettings|undefined>(() => {
             if (!optionState.dashboardSettings) return undefined;
             const dateRange = optionState.dashboardSettings.date_range;
@@ -96,58 +84,13 @@ export function mergeBaseWidgetState(
         title: computed<string>(() => optionState.title ?? state.widgetConfig?.title ?? ''),
         schemaProperties: computed<string[]>(() => {
             const initialSchemaProperties = getInitialSchemaProperties(state.widgetConfig, optionState.dashboardVariablesSchema);
-            const updatedSchemaProperties = getSchemaPropertiesByDashboardScope(initialSchemaProperties, optionState.dashboardScope ?? 'WORKSPACE');
-            if (!optionState.schemaProperties) return updatedSchemaProperties;
-            return getRefinedSchemaProperties(optionState.schemaProperties, updatedSchemaProperties, optionState.widgetOptions);
+            if (!optionState.schemaProperties) return initialSchemaProperties;
+            return getRefinedSchemaProperties(optionState.schemaProperties, initialSchemaProperties, optionState.widgetOptions);
         }),
     }) as UnwrapRef<MergedBaseWidgetState>;
 
     return state;
 }
-
-
-const getWidgetConfigByDashboardScope = (config: WidgetConfig, dashboardScope: DashboardScope): WidgetConfig => {
-    if (dashboardScope === 'DOMAIN') {
-        const extraOptionsSchema = getWidgetOptionsSchema(['filters.workspace']);
-        return {
-            ...config,
-            options_schema: {
-                properties: { ...(config.options_schema?.properties ?? {}), ...extraOptionsSchema.properties },
-                order: ['filters.workspace', ...(config.options_schema?.order ?? [])],
-            },
-        };
-    }
-    if (dashboardScope === 'PROJECT') {
-        const extraOptionsSchema = getWidgetOptionsSchema([['filters.project', { fixed: true, optional: false, readonly: true }]]);
-        return {
-            ...config,
-            options_schema: {
-                properties: { ...(config.options_schema?.properties ?? {}), ...extraOptionsSchema.properties },
-                order: ['filters.project', ...(config.options_schema?.order ?? [])],
-            },
-        };
-    }
-    return config;
-};
-const getInheritOptionsByDashboardScope = (inheritOptions: InheritOptions, dashboardScope: DashboardScope): InheritOptions => {
-    if (dashboardScope === 'DOMAIN') {
-        return { ...inheritOptions, 'filters.workspace': { enabled: true, variable_key: 'workspace' } };
-    }
-    if (dashboardScope === 'PROJECT') {
-        return { ...inheritOptions, 'filters.project': { enabled: true, variable_key: 'project' } };
-    }
-    return inheritOptions;
-};
-
-const getSchemaPropertiesByDashboardScope = (schemaProperties: string[], dashboardScope: DashboardScope): string[] => {
-    if (dashboardScope === 'DOMAIN') {
-        return union(['filters.workspace'], schemaProperties);
-    }
-    if (dashboardScope === 'PROJECT') {
-        return union(['filters.project'], schemaProperties);
-    }
-    return schemaProperties;
-};
